Predictor/requirements.txt
PyYAML==6.0.1
autopep8==2.0.4
numpy==1.26.4
pandas==2.2.1
matplotlib==3.8.3
xgboost==2.0.3
scikit-learn==1.4.1.post1